split model into 2 meshes to solve indices overflow bug

+++ src/world.cpp
@@ -7,7 +7,7 @@ void load_chunks() {
     for (int i = 0; i < 16; i++) {
         for (int j = 0; j < 16; j++) {
             for (int k = 0; k < 16; k++) {
-                chunk.add_block({i, k, j}, BLOCK::GRASS);
+                chunk.add_block({(float) i, (float) k, (float) j}, BLOCK::GRASS);
             }
         }
     }
